requirements.txt
pandas>=2.0
matplotlib>=3.7
numpy>=1.24
PyYAML>=6.0
Pillow>=10.0
